refactor: rename group.ts to group.svelte.ts for rune compatibility
Updated file extension from .ts to .svelte.ts to properly handle Svelte 5 rune types used in alertState. Also updated import paths in CharConfig.svelte and index.svelte.ts to reference the new file location.
